+++ Changes.md
@@ -4,7 +4,9 @@
 Breaking Changes
 ----------------
 
-- CyclesOptions : - Removed all texture cache options. These had never been exposed in the UI because this never became an offical Cycles feature.
+- CyclesOptions :
+  - Removed `useFrameAsSeed` plug. The frame is now automatically used as the seed if `seed` is not set.
+  - Removed all texture cache options. These had never been exposed in the UI because this never became an offical Cycles feature.
 
 1.4.0.0b1 (relative to 1.3.x.x)
 =========